+++ packages/terra-avatar/tests/wdio/avatar-common-spec.js
@@ -1,63 +1,55 @@
-describe('Avatar', () => {
+Terra.describeViewports('Avatar', ['huge'], () => {
   describe('Icon Avatar', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/user-avatar'));
 
-    Terra.it.isAccessible({ selector: '#user-avatar' });
-    Terra.it.matchesScreenshot({ selector: '#user-avatar' });
+    Terra.it.validatesElement({ selector: '#user-avatar' });
   });
 
   describe('Image Avatar', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/image-avatar'));
 
-    Terra.it.isAccessible({ selector: '#image-avatar' });
-    Terra.it.matchesScreenshot({ selector: '#image-avatar' });
+    Terra.it.validatesElement({ selector: '#image-avatar' });
   });
 
   describe('Invalid Image Avatar', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/invalid-image-avatar'));
 
-    Terra.it.isAccessible({ selector: '#invalid-image-avatar' });
-    Terra.it.matchesScreenshot({ selector: '#invalid-image-avatar' });
+    Terra.it.validatesElement({ selector: '#invalid-image-avatar' });
   });
 
   describe('Image Avatar Spacing', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/image-gap'));
-    Terra.it.matchesScreenshot();
+    Terra.it.validatesElement();
   });
 
   describe('One Initial Avatar', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/one-initial-avatar'));
 
-    Terra.it.isAccessible({ selector: '#one-initial-avatar' });
-    Terra.it.matchesScreenshot({ selector: '#one-initial-avatar' });
+    Terra.it.validatesElement({ selector: '#one-initial-avatar' });
   });
 
   describe('Two Initials Avatar', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/two-initials-avatar'));
 
-    Terra.it.isAccessible({ selector: '#two-initials-avatar' });
-    Terra.it.matchesScreenshot({ selector: '#two-initials-avatar' });
+    Terra.it.validatesElement({ selector: '#two-initials-avatar' });
   });
 
   describe('Deceased Avatar', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/is-deceased-avatar'));
 
-    Terra.it.isAccessible({ selector: '#is-deceased-avatar' });
-    Terra.it.matchesScreenshot({ selector: '#is-deceased-avatar' });
+    Terra.it.validatesElement({ selector: '#is-deceased-avatar' });
   });
 
   describe('Deceased Initials Avatar', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/is-deceased-initials-avatar'));
 
-    Terra.it.isAccessible({ selector: '#is-deceased-initials-avatar' });
-    Terra.it.matchesScreenshot({ selector: '#is-deceased-initials-avatar' });
+    Terra.it.validatesElement({ selector: '#is-deceased-initials-avatar' });
   });
 
   describe('Deceased Image Avatar', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/is-deceased-image-avatar'));
 
-    Terra.it.isAccessible({ selector: '#is-deceased-image-avatar' });
-    Terra.it.matchesScreenshot({ selector: '#is-deceased-image-avatar' });
+    Terra.it.validatesElement({ selector: '#is-deceased-image-avatar' });
   });
 });
 
@@ -65,22 +57,19 @@ describe('Facility', () => {
   describe('Icon Facility', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/facility/default-facility'));
 
-    Terra.it.isAccessible({ selector: '#default-facility' });
-    Terra.it.matchesScreenshot({ selector: '#default-facility' });
+    Terra.it.validatesElement({ selector: '#default-facility' });
   });
 
   describe('Image Facility', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/facility/image-facility'));
 
-    Terra.it.isAccessible({ selector: '#image-facility' });
-    Terra.it.matchesScreenshot({ selector: '#image-facility' });
+    Terra.it.validatesElement({ selector: '#image-facility' });
   });
 
   describe('Invalid Image Facility', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/facility/invalid-image-facility'));
 
-    Terra.it.isAccessible({ selector: '#invalid-image-facility' });
-    Terra.it.matchesScreenshot({ selector: '#invalid-image-facility' });
+    Terra.it.validatesElement({ selector: '#invalid-image-facility' });
   });
 });
 
@@ -88,7 +77,6 @@ describe('Shared User', () => {
   describe('Icon Shared User', () => {
     before(() => browser.url('/#/raw/tests/terra-avatar/avatar/shared-user/default-shared-user'));
 
-    Terra.it.isAccessible({ selector: '#shared-user' });
-    Terra.it.matchesScreenshot({ selector: '#shared-user' });
+    Terra.it.validatesElement({ selector: '#shared-user' });
   });
 });