+++ packages/terra-paginator/src/terra-dev-site/doc/example/ProgressivePaginatorExample.jsx
@@ -2,7 +2,10 @@ import React from 'react';
 import Dialog from 'terra-dialog';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import ProgressivePaginator from 'terra-paginator/lib/ProgressivePaginator';
+import classNames from 'classnames/bind';
+import styles from './PaginatorExampleCommon.module.scss';
 
+const cx = classNames.bind(styles);
 const totalCount = 450;
 
 const fillArray = (value, len) => {
@@ -46,7 +49,7 @@ class ProgressivePaginatorExample extends React.Component {
 
   render() {
     return (
-      <div style={{ height: '500px' }}>
+      <div className={cx('paginator-wrapper')}>
         <Dialog
           header={(
             <h1>