+++ packages/terra-status-view/src/terra-dev-site/test/status-view/StatusViewDefault.test.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import StatusView from '../../../StatusView';
-
-export default () => (
-  <div>
-    <StatusView id="statusView" />
-  </div>
-);