+++ packages/terra-text/src/terra-dev-site/test/text/TextLineHeight.test.jsx
@@ -1,15 +1,19 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import Text from '../../../Text';
+import styles from './TextTestCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 const TextLineHeight = () => (
   <div>
-    <div style={{ fontSize: '100px' }}>
+    <div className={cx('content-font-size-large')}>
       Parent font size: 100px
       <Text fontSize={16}>
         Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla at ex tellus. Duis eget maximus ante, vitae bibendum libero. Integer.
       </Text>
     </div>
-    <div style={{ 'font-size': '8px' }}>
+    <div className={cx('content-font-size-small')}>
       Parent font size: 8px
       <Text id="lineHeight" fontSize={16}>
         Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla at ex tellus. Duis eget maximus ante, vitae bibendum libero. Integer.