+++ packages/terra-arrange/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports
+
+3.12.0 - (June 28, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 3.11.0 - (June 18, 2019)
 ------------------