+++ packages/terra-form-fieldset/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+2.16.0 - (June 28, 2019)
+------------------
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
 2.15.0 - (June 18, 2019)
 ------------------
 ### Removed