+++ packages/terra-i18n/CHANGELOG.md
@@ -3,8 +3,14 @@ Changelog
 
 Unreleased
 ----------
+
+4.7.0 - (June 28, 2019)
+------------------
 ### Added
 * Added more information to upgrade guide around Jest testing
+
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
 
 4.6.0 - (June 18, 2019)
 ------------------