+++ packages/terra-list/CHANGELOG.md
@@ -3,9 +3,13 @@ Changelog
 
 Unreleased
 ----------
+
+4.9.0 - (June 28, 2019)
+------------------
 ### Changed
 * Fixed accessibility issue in list item tests
 * Update mixin import to work better with codesandbox.io
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
 
 4.8.0 - (June 18, 2019)
 ------------------