+++ packages/terra-signature/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.13.0 - (June 28, 2019)
+------------------
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
 2.12.0 - (June 11, 2019)
 ------------------
 ### Removed