+++ packages/terra-text/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+4.8.0 - (June 28, 2019)
+------------------
+### Changed
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
+
 4.7.0 - (June 18, 2019)
 ------------------
 ### Changed